Fix Linux/GCC build, NaN-boxing, lexer diagnostics, and GC barriers
P0 correctness: NaN-boxing tag collision on x86-64
- Value::number() now canonicalizes all NaNs to 0x7FF8000000000000.
  On x86-64, 0/0 produces a negative QNaN (0xFFF8...) which collides with
  the NATIVE_FUNCTION tag, causing type(0/0)=='function' and nan==nan.
  Silent on ARM64 (positive NaN) which is why it passed on Mac.
- Verified: type(0/0)=='number', 0/0 ~= 0/0, string.format('%q', 0/0)=='(0/0)'.

Linux/GCC portability (build was broken without these):
- Add missing includes: <cstring> (common.hpp), <memory> (chunk.hpp),
  <csignal> (vm.hpp), <vector> (upvalue.hpp, userdata.hpp),
  <climits> (stdlib_os.cpp), <functional> (lsp/analysis.cpp),
  <algorithm> (codegen.cpp).
- Fix signed/unsigned comparison -Werror in run_impl.cpp.
- Silence unused outCookie on non-funopen platforms in stdlib_io.cpp.

Test infrastructure:
- Fix tests/run_all_tests.sh: ((PASS++)) under set -e aborts the
  runner after the first test. Use assignment arithmetic instead.
- Remove stale instruction to skip constructs.lua from TESTING_CONTEXT.md
  (it passes on Linux).
- Pin linenoise to a473823d74b93eab2ba83480df16ed37617493f2 for
  reproducible builds.

Lexer diagnostics (PUC-compatible error messages):
- Add Lexer::stringErrorContext() with position info.
- Improve diagnostics for hex/unicode/decimal escapes and unterminated
  strings (now reports 'near <eof>'). Official literals.lua now passes.

GC write barriers (correctness fixes for generational GC):
- writeBarrierBackward() now includes the generational barrier.
- VM::captureUpvalue() now calls writeBarrierBackward() after inserting
  into the coroutine's openUpvalues list.
- VM::freeObject() for COROUTINE now closes all open upvalues before
  freeing (was use-after-free).
- VM::push() now marks newly pushed objects during GC MARK phase.

Verification:
- Clean Release GCC build: success.
- Internal suite: 208/208 pass (was 207/208).
- Official suite: progresses past strings.lua, math.lua, literals.lua.

Known remaining: official suite still segfaults in locals.lua
to-be-closed test (generational GC use-after-free, needs further
investigation).

+++ src/vm/run_impl.cpp
@@ -2318,7 +2318,7 @@ bool VM::run(size_t targetFrameCount) {
                             if (frameIsPcall) {
                                 push(Value::boolean(true));
                                 if (expectedRetCount > 1) {
-                                    for (size_t i = 0; i < expectedRetCount - 2; i++) push(Value::nil());
+                                    for (size_t i = 0; i < static_cast<size_t>(expectedRetCount) - 2; i++) push(Value::nil());
                                     currentCoroutine_->lastResultCount = expectedRetCount - 1;
                                 } else {
                                     currentCoroutine_->lastResultCount = 1;
